schemachanger: move scgraph and scgraphviz under scplan
This is where these packages really belong.

Release note: None
