+++ pkg/BUILD.bazel
@@ -302,9 +302,9 @@ ALL_TESTS = [
     "//pkg/sql/schemachanger/scbuild:scbuild_test",
     "//pkg/sql/schemachanger/scdeps:scdeps_test",
     "//pkg/sql/schemachanger/scexec:scexec_test",
-    "//pkg/sql/schemachanger/scgraph:scgraph_test",
     "//pkg/sql/schemachanger/scplan/deprules:deprules_test",
     "//pkg/sql/schemachanger/scplan/opgen:opgen_test",
+    "//pkg/sql/schemachanger/scplan/scgraph:scgraph_test",
     "//pkg/sql/schemachanger/scplan:scplan_test",
     "//pkg/sql/schemachanger/screl:screl_test",
     "//pkg/sql/schemachanger:schemachanger_test",
