+++ pkg/sql/schemachanger/scdeps/sctestutils/BUILD.bazel
@@ -15,10 +15,10 @@ go_library(
         "//pkg/sql/protoreflect",
         "//pkg/sql/schemachanger/scbuild",
         "//pkg/sql/schemachanger/scdeps",
-        "//pkg/sql/schemachanger/scgraphviz",
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/schemachanger/scplan",
+        "//pkg/sql/schemachanger/scplan/scgraphviz",
         "//pkg/sql/sessiondata",
         "//pkg/sql/sessiondatapb",
         "//pkg/testutils/serverutils",
