+++ pkg/sql/schemachanger/scplan/BUILD.bazel
@@ -7,11 +7,11 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/jobs/jobspb",
-        "//pkg/sql/schemachanger/scgraph",
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/schemachanger/scplan/deprules",
         "//pkg/sql/schemachanger/scplan/opgen",
+        "//pkg/sql/schemachanger/scplan/scgraph",
         "//pkg/sql/schemachanger/scplan/scopt",
         "//pkg/sql/schemachanger/scplan/scstage",
         "@com_github_cockroachdb_errors//:errors",
@@ -37,10 +37,10 @@ go_test(
         "//pkg/sql/schemachanger/scbuild",
         "//pkg/sql/schemachanger/scdeps/sctestutils",
         "//pkg/sql/schemachanger/scerrors",
-        "//pkg/sql/schemachanger/scgraph",
-        "//pkg/sql/schemachanger/scgraphviz",
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
+        "//pkg/sql/schemachanger/scplan/scgraph",
+        "//pkg/sql/schemachanger/scplan/scgraphviz",
         "//pkg/sql/schemachanger/scplan/scstage",
         "//pkg/sql/schemachanger/screl",
         "//pkg/sql/sem/tree",
