+++ pkg/sql/schemachanger/scplan/deprules/BUILD.bazel
@@ -12,8 +12,8 @@ go_library(
     deps = [
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/schemachanger/rel",
-        "//pkg/sql/schemachanger/scgraph",
         "//pkg/sql/schemachanger/scpb",
+        "//pkg/sql/schemachanger/scplan/scgraph",
         "//pkg/sql/schemachanger/screl",
         "@com_github_cockroachdb_errors//:errors",
     ],
