+++ pkg/sql/schemachanger/scplan/opgen/BUILD.bazel
@@ -47,9 +47,9 @@ go_library(
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/schemachanger/rel",
-        "//pkg/sql/schemachanger/scgraph",
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
+        "//pkg/sql/schemachanger/scplan/scgraph",
         "//pkg/sql/schemachanger/screl",
         "//pkg/util/protoutil",
         "@com_github_cockroachdb_errors//:errors",
