+++ pkg/sql/schemachanger/scplan/opgen/op_gen.go
@@ -11,9 +11,9 @@
 package opgen
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scgraph"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scop"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scpb"
+	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scplan/scgraph"
 )
 
 type registry struct {
