+++ pkg/sql/schemachanger/scplan/scgraphviz/BUILD.bazel
@@ -3,13 +3,13 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "scgraphviz",
     srcs = ["graphviz.go"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scgraphviz",
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scplan/scgraphviz",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/sql/schemachanger/scgraph",
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/schemachanger/scplan",
+        "//pkg/sql/schemachanger/scplan/scgraph",
         "//pkg/util/protoutil",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_emicklei_dot//:dot",
