+++ pkg/sql/schemachanger/scplan/scstage/BUILD.bazel
@@ -11,9 +11,9 @@ go_library(
     deps = [
         "//pkg/jobs/jobspb",
         "//pkg/sql/catalog/descpb",
-        "//pkg/sql/schemachanger/scgraph",
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
+        "//pkg/sql/schemachanger/scplan/scgraph",
         "//pkg/sql/schemachanger/screl",
         "//pkg/util/iterutil",
         "@com_github_cockroachdb_errors//:errors",
