+++ pkg/sql/schemachanger/scrun/BUILD.bazel
@@ -13,10 +13,10 @@ go_library(
         "//pkg/jobs/jobspb",
         "//pkg/settings/cluster",
         "//pkg/sql/schemachanger/scexec",
-        "//pkg/sql/schemachanger/scgraphviz",
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/schemachanger/scplan",
+        "//pkg/sql/schemachanger/scplan/scgraphviz",
         "//pkg/sql/schemachanger/scplan/scstage",
         "//pkg/util/log/logcrash",
         "@com_github_cockroachdb_errors//:errors",
